Move preview-specific files into new package
